requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
python-jose[cryptography]
slowapi
structlog
cachetools
